Address comments in #279
