+++ CMakeLists.txt
@@ -279,6 +279,11 @@ set(NETKET_SOURCES
     Sources/Sampler/vmc_sampling.cc
     Sources/Sampler/metropolis_hastings.cc
     Sources/Sampler/metropolis_hastings_pt.cc
+    Sources/Sampler/custom_local_kernel.cc
+    Sources/Sampler/exchange_kernel.cc
+    Sources/Sampler/hamiltonian_kernel.cc
+    Sources/Sampler/hop_kernel.cc
+    Sources/Sampler/local_kernel.cc
     Sources/Stats/mc_stats.cc
     Sources/Stats/py_stats.cc
     Sources/Optimizer/stochastic_reconfiguration.cc
